Version 23.010
modified:   document-groups/37151457af40ee706cc23de4a11e7ebacafd0c10/document-types/abb85feaa97ca39b1bdb0e8a29359f1995fdcc8d/update.php
modified:   document-groups/a3f5d7ead80d4e6cb644ec158a13f3a89a9a0622/document-types/01fe006d1aca8e85fc140fb642bb200ed6e31596/update.php
-   Fixed a bug that would delete the new thumbnail image file if its filename was the same as the previous thumbnail image.

+++ document-groups/37151457af40ee706cc23de4a11e7ebacafd0c10/document-types/abb85feaa97ca39b1bdb0e8a29359f1995fdcc8d/update.php
@@ -155,7 +155,8 @@ if (isset($_FILES['image-file']))
     {
         $filename = $archive->absoluteDataDirectory() . '/' . $thumbnailImageBasename;
 
-        if (is_file($filename))
+        if ($filename != $thumbnailImageFilename &&
+            is_file($filename))
         {
             unlink($filename);
         }
