+++ document-groups/a3f5d7ead80d4e6cb644ec158a13f3a89a9a0622/document-types/01fe006d1aca8e85fc140fb642bb200ed6e31596/update.php
@@ -169,7 +169,8 @@ if (isset($_FILES['image-file']))
     {
         $filename = $archive->absoluteDataDirectory() . '/' . $thumbnailImageBasename;
 
-        if (is_file($filename))
+        if ($filename != $thumbnailImageFilename &&
+            is_file($filename))
         {
             unlink($filename);
         }
